pyproject.toml
[project]
name = "mas-tree-of-thought"
version = "0.1.0"
description = "A multi-agent system that implements Tree of Thoughts (ToT) methodology, where multiple agents collaborate under a coordinator to solve complex problems through structured reasoning and LLM-powered evaluation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.5.0",
    "litellm>=1.66.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0"
]

[build-system]
requires = ["uv_build>=0.6.9"]
build-backend = "uv_build"
